requirements.txt
Django>=4.2
djangorestframework>=3.14
psycopg2-binary>=2.9
django-cors-headers>=3.13
gunicorn>=20.1
binance-connector>=3.12.0
yfinance
celery